+++ RELEASES.md
@@ -17,6 +17,7 @@ Upcoming
 - Introduction of the HashPipe
 - Replace configurations.<config-name>.throwConfigNotFoundException property with configurations.<config-name>.configNotFoundReportLevel (possible values: ERROR (default), WARN, INFO, DEBUG)
 - Make ibisstore checks capable for MS SQL (next to Oracle)
+- Add throwJavaListenerNotFoundException attribute to IbisLocalSender